// web/src/lib/api/generated/models/meetingContextRequest.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { ExploreSelection } from "./exploreSelection";
import type { MeetingContextRequestFormat } from "./meetingContextRequestFormat";

export type MeetingContextRequest =
  | ({
      /**
       * @minItems 1
       * @maxItems 100
       * @items.minimum 1
       * @items.maximum 9007199254740991
       */
      message_ids: number[];
    } & {
      format?: MeetingContextRequestFormat;
      include_transcript?: boolean;
      /**
       * @minimum 4096
       * @maximum 1048576
       */
      max_bytes?: number;
      /**
       * @minItems 1
       * @maxItems 100
       * @items.minimum 1
       * @items.maximum 9007199254740991
       */
      message_ids?: number[];
      selection?: ExploreSelection;
    })
  | ({
      selection: ExploreSelection;
    } & {
      format?: MeetingContextRequestFormat;
      include_transcript?: boolean;
      /**
       * @minimum 4096
       * @maximum 1048576
       */
      max_bytes?: number;
      /**
       * @minItems 1
       * @maxItems 100
       * @items.minimum 1
       * @items.maximum 9007199254740991
       */
      message_ids?: number[];
      selection?: ExploreSelection;
    });